Move login to auth module
