+++ inductiva/__init__.py
@@ -25,6 +25,7 @@ from . import users
 from . import logs
 from . import api
 from . import events
+from . import auth
 from .templating import TemplateManager
 
 logs.setup(getattr(logging, os.environ.get("INDUCTIVA_LOG_LEVEL", "INFO")))
