+++ inductiva/auth/__init__.py
@@ -1,0 +1,2 @@
+#pylint: disable=missing-module-docstring
+from inductiva.auth.methods import login, logout
