+++ inductiva/users/__init__.py
@@ -1,2 +1,2 @@
 #pylint: disable=missing-module-docstring
-from inductiva.users.methods import get_quotas, get_info, get_costs, login, logout
+from inductiva.users.methods import get_quotas, get_info, get_costs
