+++ inductiva/users/methods.py
@@ -82,22 +82,3 @@ def get_costs(start_year: int,
 
         request = api_instance.get_user_costs(query_params=query_params)
     return request.body["costs"]
-
-
-def login(private: bool = False) -> None:
-    """Logs the user in to the Inductiva platform.
-
-    This function handles user login by prompting the user to log in via the
-      command line.
-
-    Args:
-        private:  If True, no api_key will be printed to the console.
-    """
-
-    args = argparse.Namespace(private=private)
-    login_cmd(args)
-
-
-def logout():
-    """Logs the user out by removing the stored API key."""
-    logout_cmd(None)
